refactor(core): remove core.lua and adopt services as shared primitives
test: split legacy core_spec into service-focused specs
chore(docs): align handlers/services AGENTS boundaries
