+++ lua/opencode/commands/handlers/AGENTS.md
@@ -1,12 +1,22 @@
 # AGENTS.md (handlers)
 
-This directory owns domain actions and command definitions.
+This directory owns command-facing action adapters and command definitions.
+
+One-line positioning:
+- `handlers/` = **command-entry adapters**
+- `services/` = **cross-entry reusable business primitives**
 
 ## Scope
 
-- `M.actions`: domain operations
+- `M.actions`: command-facing action adapters
 - `M.command_defs`: command-facing definitions (desc/completions/execute)
 - No command pipeline logic here
+
+## Relation with services
+
+- Handlers should call `services/*` when logic is shared by command/UI/quick_chat entries.
+- Handlers should not become the only place for reusable business logic.
+- If an action is needed outside command entry paths, move/keep it in `services/`.
 
 ## Hard Invariants
 
@@ -15,12 +25,24 @@ This directory owns domain actions and command definitions.
 - Handlers do not decide hook routing
 - Keep action behavior identical across entry points
 - Handlers must not introduce any new bind/execute entry symbols (`*.run`, `bind_*`, or dispatch wrappers)
+- Prefer `services/*` over direct new `opencode.session` / `opencode.api` requires in handlers
 
 ## Structure Guidance
 
 - Keep `actions` and `command_defs` aligned by domain
 - Keep keymap compatibility aliases explicit and grouped
 - Avoid duplicating command validation already guaranteed by parse schema
+
+## Current boundary debt (file-level TODO)
+
+The following handler files still directly require `opencode.session` and should be routed via services APIs.
+
+- [ ] `lua/opencode/commands/handlers/diff.lua` -> `opencode.session`
+- [ ] `lua/opencode/commands/handlers/session.lua` -> `opencode.session`
+
+Sync rule:
+- keep this list aligned with `lua/opencode/services/AGENTS.md`
+- remove an item only after code + tests pass
 
 ## Editing Rules
 
@@ -46,12 +68,14 @@ This directory owns domain actions and command definitions.
 - Is `command_defs` declarative and minimal?
 - Are aliases grouped and obvious?
 - Did we avoid reintroducing duplicate argument validation paths?
+- Did we add any new direct `session/api` requires in handlers?
 
 ## Reject Conditions
 
 - Any direct call to `dispatch.execute` from handlers
 - Any parse/hook routing logic added to handlers
 - Any new entry-style wrapper added in handlers
+- Any new direct `require('opencode.session')` or `require('opencode.api')` in handlers without exception note
 
 ## Minimal Regression Commands
 
@@ -63,6 +87,6 @@ This directory owns domain actions and command definitions.
 ## Entry Notes For New Agents
 
 - Start from the domain file you are touching (`window/session/diff/workflow/surface/agent/permission`), then verify invariants against `commands/dispatch.lua`.
-- Treat handlers as **domain behavior + command definition only**.
+- Treat handlers as **command adaptation + command definition only**.
 - If you feel the need to touch parse/dispatch from handlers, stop and move that change to the command infrastructure layer.
 - Keep compatibility aliases explicit, local, and justified inline.
